Refatora autenticação JWT e centraliza políticas de auth
Refatora a configuração de autenticação JWT para uma extensão dedicada (AddJwtAuthentication) e centraliza as políticas de autorização em AuthorizationPolicies. Ajusta a busca do UUID no AuthController e padroniza o uso das variáveis de ambiente JWT_KEY, JWT_ISSUER e JWT_AUDIENCE no TokenService.
